Split YouTube into common traits
