+++ uploader/src/main/scala/com/gu/media/UploaderLambda.scala
@@ -3,12 +3,13 @@ package com.gu.media
 import com.gu.media.UploaderLambda.{Input, Output}
 import com.gu.media.lambda.LambdaWithParams
 import com.gu.media.logging.Logging
+import com.gu.media.youtube.YouTube
 import org.cvogt.play.json.Jsonx
 import play.api.libs.json.Format
 
 class UploaderLambda extends LambdaWithParams[Input, Output] with Logging {
   override def handleRequest(input: Input): Output = {
-    val youTube = YouTubeClient(config)
+    val youTube = new YouTube(config)
     val output = s"allowedChannels: ${youTube.allowedChannels.mkString(",")}. result: ${input.string * input.integer}"
 
     log.info(output)
